% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_shp.R
\name{cat_shp}
\alias{cat_shp}
\title{Catchment Shape}
\usage{
cat_shp(netcat)
}
\arguments{
\item{netcat}{output from \code{net_delin}}
}
\value{
a \code{data.frame}: \code{group.comid} of network root node; \code{net.comid}
  network flowline comid; and basin.len which is basin lenght
}
\description{
Calculates metrics related to catchment \code{Value}
}
\details{
Requires netcat object (see \code{\link{net_cat}})
}
\examples{
g <- cat_shp(netcat = c, vpu = "01", nhdplus_path = getwd())

}
