% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_bfw.R
\name{net_bfw}
\alias{net_bfw}
\title{Bankfull Width Estimate}
\usage{
net_bfw(netdelin, vpu, nhdplus_path)
}
\arguments{
\item{netdelin}{output from \code{net_delin}}

\item{vpu}{NHDPlusV2 Vector Processing Unit}

\item{nhdplus_path}{Directory for NHDPlusV2 files (\code{\link{net_nhdplus}})}
}
\value{
\code{data.frame}: \code{$net.id} unique identifier for stream
  network; \code{$group.comid} CoMID of stream network root; \code{M}
  Position of sampling point on COMID, as proportion of COMID from upstream
  end; \code{$AREASQKM} drainage area (km^2) scaled by M value;
  \code{est_bfw} estimated bankfull width; \code{bfw_fit} model fit for
  bankfull width - drainage area relationship
}
\description{
Estimates bankfull channel width at comid outlet (unless M-value is
specified) from drainage area using regression coefficients in Faustini et al
(2009) Geomorphology
}
\details{
Requires /NHDPlusAttributes directory (see \code{\link{net_nhdplus}})

Drainage area measures are scaled by M values
}
\examples{
# identify NHDPlusV2 COMID
a <- net_sample(nhdplus_path = getwd(), vpu = "01", ws_order = 6, n = 5)
# delineate stream network
b <- net_delin(group_comid = as.character(a[,"COMID"]), nhdplus_path = getwd(), vpu = "01")
calculate topology summary
c <- net_bfw(netdelin = b, vpu = "01", nhdplus_path = getwd())
}
