% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_calc.r
\name{net_calc}
\alias{net_calc}
\title{Network Topology}
\usage{
net_calc(netdelin, vpu, nhdplus_path)
}
\arguments{
\item{netdelin}{output from \code{net_delin}}

\item{vpu}{vector processing unit}

\item{nhdplus_path}{directory containing NHDPlusV2 \code{\link{net_nhdplus}}}
}
\value{
a \code{data.frame}: \code{COMID} of network root node; \code{vpu}
   vector processing unit; \code{maxelev} max elevation of network flowline (cm);
  \code{minelev} min elevation of network flowline (cm); \code{releif}
  (\code{maxelev} - \code{minelev}); \code{drain.dia} drainage diameter
  (longest flowpath in network); \code{head.h2o} number of terminal nodes
  (i.e. headwaters); \code{trib.jun} number of tributary junctions
  (\code{head.h2o} - 1); \code{AREASQKM} drainage area (km^2);
  \code{LENGTHKM} total lenght of network flowlines; \code{drain.den}
  drainage density (\code{LENGTHKM} / \code{AREASQKM}); \code{WS.ord}
  strahler order for root node; \code{reach.cnt} number of valley segments
  (edges); \code{diver.cnt} count of divergent flow paths
}
\description{
Calculates metrics related to stream network topology see \code{Value}
}
\details{
Requires /NHDPlusAttributes directory (see \code{\link{net_nhdplus}})

Warning: "Check drain.dia. Incomplete headwater path 5848788 at COMID 5848656
in network 5848066 perhaps due to waterbody" may indicate that
drain.dia is flawed by the presence of a waterbody in the network
}
\examples{
g <- net_calc(netdelin = c, vpu = "01", nhdplus_path = getwd())
}
