% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_clim.r
\name{net_clim}
\alias{net_clim}
\title{Network Climate Summary}
\usage{
net_clim(nhdplus_path, vpu, netdelin)
}
\arguments{
\item{nhdplus_path}{directory containing NHDPlus
see(\code{\link{net_nhdplus}})}

\item{vpu}{the vector processing unit}

\item{netdelin}{output from \code{\link{net_delin}}}
}
\value{
\code{data.frame} containing: \code{COMID} root node of network;
 \code{vpu} vector processing unit; \code{MISSDATAA.Temp} area of missing
 temperature data; \code{TEMPVC} mean annual temperature (deg C); \code{seasonality_t}
 Coefficient of variation of mean monthly temperatures; \code{warm_mo} warmest
 month; \code{warm_mo_t} mean temperature of warmest month; \code{cold_mo}
 coldest month; \code{cold_mo_t} mean temperature of coldest month;
 \code{diff_t} temperature difference between warm and cold monthly
 temperatures; \code{warm_q_t} mean temperature of warmest quarter;
 \code{warm_q}  warmest quarter; \code{cold_q_t} mean temperature of coldest
 quarter; \code{cold_q} coldest quarter; \code{MISSDATAA.Ppt} area of missing
 precipitation data; \code{PRECIPVC} cumulative mean annual precipiration (mm);
 \code{wet_mo} wettest month; \code{wet_mo_p} cumulative mean precipitation
 of wettest month; \code{dry_mo} driest month; \code{dry_mo_p} cumulative
 mean precipitation of driest month; \code{seasonality_p} coefficient of
 vatiation of mean monthly precipitation; \code{wet_q_p} cumulaltive mean
 precipitation of wettest quarter; \code{wet_q} wettest quarter;
 \code{dry_q_p} cumulaltive mean precipitation of driest quarter; \code{dry_q}
 driest quarter; \code{dry_q_t} mean temperature of driest quarter;
 \code{wet_q_t} mean temperature of wettest quarter; \code{warm_q_p}
 cumulaltive mean precipitation of warmest quarter; \code{cold_q_p}
 cumulaltive mean precipitation of coldest quarter; \code{RUNOFFVC}
 cumulative mean annual runff (mm); \code{MAFLOWV} Vogal Estimation of mean
 annual discharge; \code{MAVELV} Vogal estimation of mean annual velocity at
 network outlet; \code{Q0001E} EROM estimation of mean mean annual
 discharge; \code{V0001E} EROM estimation of mean annual velocity
}
\description{
network scale summeries of bioclim
(\url{http://www.worldclim.org/bioclim}) variables
}
\details{
Requires "/VPUAttributeExtension" directory see(\code{\link{net_nhdplus}})
}
\examples{
f <- net_clim(netdelin = c,vpu = "01", nhdplus_path = getwd())
}
