% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_delin.r
\name{net_delin}
\alias{net_delin}
\title{Network Delineation}
\usage{
net_delin(group_comid, nhdplus_path = getwd(), vpu, M = NULL,
  snap_xy = NULL)
}
\arguments{
\item{group_comid}{a vector of NHDPlusV2 COMIDs}

\item{nhdplus_path}{Directory for NHDPlusV2 files (\code{\link{net_nhdplus}})}

\item{vpu}{NHDPlusV2 Vector Processing Unit}

\item{M}{Position of sampling point on COMID (optional). Generated from
\code{\link{net_comid}}}

\item{snap_xy}{coordinates of sampling point on COMID (optional). Generated
from \code{\link{net_comid}}}
}
\value{
Named list (\code{$Network, $Nested_COMIDs, $SF_Obj}).\code{$Network}
  stores COMID as data.frame: \code{$Network$group.comid} Root COMID for
  network; \code{$Network$net.comid} COMID's upstream of
  root;\code{$Network$vpu} Vector Processing Unit; \code{$Network$M} Position
  of sampling point on COMID, as proportion of COMID from upstream end;
  \code{$Network$net.id} unique identifier for stream network.
  \code{$Nested_COMIDs} vector of comids in same stream network.
  \code{$SF_Obj} Simple Features Object: \code{$SF_Obj$group.comid} Root
  COMID of network; \code{$SF_Obj$COMID} same as \code{$Network$net.comid}
  COMID's upstream of root; \code{$SF_Obj$VPUID} same as \code{$Network$vpu}
  Vector Processing Unit; \code{$SF_Obj$Meas} same as \code{$Network$M}
  Position of sampling point on COMID; \code{$SF_Obj$Meas} same as
  \code{$Network$net.id} unique identifier for stream network
}
\description{
Identifies all flowlines within a stream network upstream of a NHDPlusV2
COMID (see \code{\link{net_sample}} or \code{\link{net_comid}}.
}
\details{
see NHDPlusV2 Documentation
\url{http://www.horizon-systems.com/NHDPlus/NHDPlusV2_data.php})

Requires /NHDPlusAttributes and NHDSnapshot directories (see
\code{\link{net_nhdplus}})

All input COMID are delineated. Nested COMIDs occur when two input comid are
in the same network. If nested comids are unexpected or unwanted, remove from
\code{$SF_Obj} and \code{$Network} before proceeding
}
\examples{
# process sampling ponits
a <- net_comid(sample_points = z, CRS = 4269, nhdplus_path = getwd(), vpu="01", maxdist = 100)
# or select random networks
a <- net_sample(nhdplus_path = getwd(), vpu = "01", ws_order = 3, n = 5)
# delineate upstream reaches
b <- net_delin(group_comid = as.character(a[,"COMID"]), nhdplus_path = getwd(), vpu = "01")
#write stream network as shpfile
sf::write_sf(b$SF_Obj,paste(getwd(),"/network.shp"))
}
