% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_flow.R
\name{net_flow}
\alias{net_flow}
\title{Network Discharge (Q0001E) Summary}
\usage{
net_flow(nhdplus_path, vpu, netdelin)
}
\arguments{
\item{nhdplus_path}{directory containing NHDPlus
see(\code{\link{net_nhdplus}})}

\item{vpu}{the vector processing unit}

\item{netdelin}{output from \code{\link{net_delin}}}
}
\value{
\code{RUNOFFVC}cumulative mean annual runoff (mm) and EROM velociety
 (V0001E) at network outlet; Mean Annual discharge (\code{MAQ0001E}), min
 mean monthly dischagre (\code{minMMQ0001E}), maximum monthly discharge
 (\code{maxMMQ0001E}) and coeffficient of variation of mean monthly discharge
 (\code{covMMQ0001E}) were EROM estimations scaled by mvalue of sampling
 point of mean mean annual discharge;
}
\description{
Network Discharge (Q0001E) Summary
}
\examples{
f <- net_flow(netdelin = c, vpu = "01", nhdplus_path = getwd())
}
