% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_hor.r
\name{net_hort}
\alias{net_hort}
\title{Network Area, Length and Bifurcation Ratios}
\usage{
net_hort(netdelin, vpu, nhdplus_path)
}
\arguments{
\item{netdelin}{output from \code{net_delin}}

\item{vpu}{vector processing unit}

\item{nhdplus_path}{directory containing NHDPlusV2 \code{\link{net_nhdplus}}}
}
\value{
named \code{list}: \code{$topology} is \code{data.frame} with count
  of stream numbers (\code{str_num}), mean of stream lengths (\code{str_len})
  and mean of drainage areas (\code{str_area}); and \code{$Horton_Est} are
  horotn estimates \code{Rb,Rl,Ra} (bifurcation, lenght and area ratio,
  respectively). \code{Rb.rsqr, Rl.rsqr, Ra.rsqr} are R^2 values for the
  \code{lm(...)} model. \code{ohm} is the strahler order of the (network -
  1) (see \code{Details}.
}
\description{
Uses \code{lm(...)} to estimates number, length and area ratios from the
respected mean values a stream network (i.e. Horton Laws).
}
\details{
requires /NHDPlusAttributes directory (see \code{net_nhdplus})

Horton estimates are rounded down to the next lowest WS order (i.e. \code{ohm}
is the strahler order of the (network - 1)) because a given sampling COMID,
my not be representative of the whole stream order.

Divergences are not included

Check output for stream orders not exceeding order of group.comid. In some
networks Error in NHDPlus vaa \code{STREAMORDE} can effect calculations.
}
\examples{
net_hort(netdelin = c, vpu = "01", nhdplus_path = getwd())
}
