% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_lc.r
\name{net_lc}
\alias{net_lc}
\title{NLCD2011 Landcover Percentages}
\usage{
net_lc(netdelin, vpu, nhdplus_path)
}
\arguments{
\item{netdelin}{output from \code{\link{net_delin}}}

\item{vpu}{NHDPlusV2 Vector Processing Unit}

\item{nhdplus_path}{Directory for NHDPlusV2 files (\code{\link{net_nhdplus}})}
}
\value{
\code{data.frame} of NLCD2011 classification codes
  (\url{https://www.mrlc.gov/nlcd11_leg.php}). Each record is percentage of
  catchment area
}
\description{
Percent of watershed covered by NLCD2011 landcover class
}
\details{
Requires VPUAttributeExtension directory see(\code{\link{net_nhdplus}})
}
\examples{
#identify sample comid
a <- net_sample(nhdplus_path = getwd(), vpu = "01", ws_order = 6, n = 5)
#delineate stream network
b <- net_delin(group_comid = as.character(a[,"COMID"]), nhdplus_path = getwd(), vpu = "01")
#calculate NLCD2011 Landcover summaries
net_lc(netdelin = b, vpu = "01", nhdplus_path = getwd())
}
