% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_nhdplus.r
\name{net_nhdplus}
\alias{net_nhdplus}
\title{NHDPlusV2 Download}
\usage{
net_nhdplus(nhdplus_path = getwd(), download = "http", vpu = "01",
  files = c("NHDPlusAttributes", "NHDSnapshot", "NHDPlusCatchment",
  "VPUAttributeExtension", "VogelExtension", "EROMExtension"),
  zip_7 = "C:/Program Files/7-Zip")
}
\arguments{
\item{nhdplus_path}{parent directory for download}

\item{download}{transfer protocal ("http" or "ftp")}

\item{vpu}{NHDPlusV2 Vector Processing Unit}

\item{files}{NHDPlusV2 data file names. Default are used for SNT functions.
See \code{details}}

\item{zip_7}{Location of the 7-zip program}
}
\value{
NHDPlusV2 data files are downloaded to "NHDPlus" directory
}
\description{
Downloads 7-zip files from NHDPlusV2 webpage to "NHDPlus" sub-directory
}
\details{
7-zip (\url{https://www.7-zip.org/download.html}) must be installed. Note
location for \code{zip_7} arguement

see \url{http://www.horizon-systems.com/NHDPlus/NHDPlusV2_documentation.php}
for NHDPlusV2 file names and descriptions. Options for \code{file} argurment
are c("FdrFac", "FdrNull", "FilledAreas", "Hydrodem", "NEDSnapshot",
"EROMExtension", "NHDPlusAttributes", "NHDPlusBurnComponents",
"NHDPlusCatchment", "NHDSnapshotFGDB", "NHDSnapshot", "VogelExtension",
"VPUAttributeExtension", "WBDSnapshot").

All raster pprocessing units (RPU) within the vpu will be downloaded and can
take considerable time

Includes \code{\link[base]{system}} commands which may not work outside
Windows OS
}
\examples{
net_nhdplus(nhdplus_path = getwd(), download = "http", vpu = "01", files =
c("NHDPlusAttributes", "NHDSnapshot", "NHDPlusCatchment", "VPUAttributeExtension",
"VogelExtension", "EROMExtension"), zip_7 = "C:/Program Files/7-Zip")
}
