% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_position.r
\name{net_posit}
\alias{net_posit}
\title{Network Poisition}
\usage{
net_posit(netdelin, nhdplus_path, vpu)
}
\arguments{
\item{netdelin}{output from \code{net_delin}}

\item{nhdplus_path}{directory for NHDPlusV2 files}

\item{vpu}{NHDPlusV2 vector Processing Unit}
}
\value{
named \code{list()} of distance matrix (sqKM).  Names are To/From comids with net.id appended
}
\description{
Calculates distances matrix for each nested (flow connected) network
}
\details{
NHDSnapshot and NHDPlusAttributes are required NHDlusV2 files (see
\code{\link{net_nhdplus}})

Adjusts distances to account for M values
}
\examples{
#read example locations from VPU 11
ExLoc <- read.csv("Sample_Locations.csv")
# reorder and rename location data.frame
ExLoc <- ExLoc[,c("SiteName","W","N")]
names(ExLoc) <- c("SITE_ID","X","Y")
#find nearest NHDPlusV2 COMID
sam_pts <- net_comid(sample_points = ExLoc, CRS = 4269,
                     nhdplus_path = getwd(), vpu = 11, maxdist = 1)
b <- net_delin(group_comid = as.character(sam_pts[,"COMID"]), nhdplus_path = getwd(), vpu = "11")
c <- net_posit(netdelin = b, nhdplus_path = getwd(), vpu = "11")
}
