% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_sample.r
\name{net_sample}
\alias{net_sample}
\title{Random Stream Network Roots}
\usage{
net_sample(nhdplus_path, vpu, ws_order, n)
}
\arguments{
\item{nhdplus_path}{character specfiying the directory for downloaded files
\code{\link{net_nhdplus}}}

\item{vpu}{the NHDPlus V2 vector processing unit}

\item{ws_order}{the strahler order of selected COMID's}

\item{n}{number of COMID's selected}
}
\value{
vector of COMID's of specified Strahler order
}
\description{
Extracts random COMID's of a specified Strahler order
from a vector processing unit. Can serve as root node
for steam network (see \code{\link{net_delin}}).
}
\details{
".\NHDPlus" must be a subdirectory of \code{nhdplus_path}.

NHDPlusAttributes and NHDSnapshot are required NHDlus downloads (see \code{\link{net_nhdplus}})
}
\examples{
net_sample(nhdplus_path=getwd(),vpu = "01", ws_order = 6, n = 5)
}
