% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_delin.r
\name{net_segid}
\alias{net_segid}
\title{Identify Network Segments}
\usage{
net_segid(netdelin, nhdplus_path, vpu)
}
\arguments{
\item{netdelin}{output from \code{\link{net_delin}}}

\item{nhdplus_path}{nhdplus_path directory for downloaded NHDPlus files
(\code{\link{net_nhdplus}})}

\item{vpu}{vector processing unit}
}
\value{
modifies \code{link{net_delin}} with additional field \code{seg.id}.
}
\description{
modifies \code{\link{net_delin}} by creating an index (\code{seg.id}) for
multiple comids occuring between confluences
}
\details{
\code{seg.id} are arbitrarily assigned to network segments (i.e. there is no
up/down stream ordering to seg.id with a network)

requires /NHDPlusAttributes directory (see \code{\link{net_nhdplus}})

optional with respect to other functions within StreamNetworkTools
}
\examples{
net_segid(netdelin = b, nhdplus_path = getwd(), vpu = "01")
}
