% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_sinu.r
\name{net_sinu}
\alias{net_sinu}
\title{Sinuosity and Slope}
\usage{
net_sinu(netdelin, nhdplus_path, vpu)
}
\arguments{
\item{netdelin}{output from \code{net_delin}}

\item{nhdplus_path}{Directory containing NHDPlusV2 \code{\link{net_nhdplus}}}

\item{vpu}{NHDPlusV2 vector processing unit}
}
\value{
\code{data.frame}: \code{$net.comid} upstream reaches from
  group.comid; \code{$group.comid} network root COMID; \code{$tot.len} length
  of reach; \code{$str.len} is straight line length of reach;
  \code{$sinuosity} \code{tot.len} / \code{str.len}; \code{$MaxElevSM}
  maximum elevation of reach; \code{$MinElevSM} minimum elevtion of reach;
  \code{SlopeNHDPlus} slope of reach
}
\description{
Calculates sinuosity and slope for each reach within a NHDPlusV2 network
}
\details{
Can \code{aggregate(...)} reach values to network-scale using group by
\code{$net.id}

\code{$MaxElevSM, $MinElevSM $SlopeNHDPlus} are values at the outlet of
group.comid. \code{tot.len, str.len, sinuosity} are scaled by M value

warning message: In st_cast.sf(netdelin_Prj, "LINESTRING") : repeating
attributes for all sub-geometries for which they may not be constant is from
changing MULTILINE to LINESTRING geometry
}
\examples{
# identify NHDPlusV2 COMID
a <- net_sample(nhdplus_path = getwd(), vpu = "01", ws_order = 6, n = 5)
# delineate stream network
b <- net_delin(group_comid = as.character(a[,"COMID"]), nhdplus_path = getwd(), vpu = "01")
# calculate sinuosity and slope
c <- net_sinu(netdelin = b, vpu = "01", nhdplus_path = getwd())
}
