% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_strmcat.R
\name{net_strmcat}
\alias{net_strmcat}
\title{Network StreamCat Data}
\usage{
net_strmcat(netdelin, strmcat_path = getwd(), Class = c("Disturbance",
  "Natural"), Scale = c("Watershed", "Riparian", "Stream segment"),
  Metric.Type = c("Agriculture", "Bio", "Climate", "Disturbance",
  "Hydrology", "Infrastructure", "Land Cover", "Land Use", "Lithology",
  "Mines", "Pollution", "Riparian", "Soil", "Topography", "Urban",
  "Water quality", "Wetness"), vpu = "01")
}
\arguments{
\item{netdelin}{see \code{net_delin}}

\item{strmcat_path}{directory containing StreamCat data}

\item{Class}{StreamCat metric class default \code{c("Disturbance",
"Natural")}}

\item{Scale}{StreamCat metric scale. Default \code{c("Watershed", "Riparian",
"Stream segment")}}

\item{Metric.Type}{StreamCat metric type. Default
\code{c("Ag/Urban","Agriculture","Bio","Climate",
"Disturbance","Hydrology", "Infrastructure", "Land Cover", "Land Use",
"Lithology", "Mines", "Pollution", "Riparian", "Soil", "Topography",
"Urban", "Water quality", "Wetness"), vpu = "01")}}

\item{vpu}{NHDPlusV2 Vector Processing Unit (i.e. Region)}
}
\value{
\code{data.frame} of streamCat variables

  function still underdevelopment  Need to imporve crosswalk table
}
\description{
Extracts StreamCat variables for network
}
\examples{
strmcat_download(strmcat_path = getwd(), Class = c("Disturbance"), Scale =
c("Watershed", "Riparian", "Stream segment"), Metric.Type =
c("Infrastructure"), vpu = "01")
net_comid()
net_strmcat(netdelin = )
}
