% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strmcat_download.R
\name{strmcat_download}
\alias{strmcat_download}
\title{StreamCat Region Download}
\usage{
strmcat_download(strmcat_path = getwd(), Class = c("Disturbance",
  "Natural"), Scale = c("Watershed", "Riparian", "Stream segment"),
  Metric.Type = c("Agriculture", "Bio", "Climate", "Disturbance",
  "Hydrology", "Infrastructure", "Land Cover", "Land Use", "Lithology",
  "Mines", "Pollution", "Riparian", "Soil", "Topography", "Urban",
  "Water quality", "Wetness"), vpu = "01")
}
\arguments{
\item{strmcat_path}{parent directory for download}

\item{Class}{StreamCat metric class default \code{c("Disturbance",
"Natural")}}

\item{Scale}{StreamCat metric scale. Default \code{c("Watershed", "Riparian",
"Stream segment")}}

\item{Metric.Type}{StreamCat metric type. Default
\code{c("Ag/Urban","Agriculture","Bio","Climate",
"Disturbance","Hydrology", "Infrastructure", "Land Cover", "Land Use",
"Lithology", "Mines", "Pollution", "Riparian", "Soil", "Topography",
"Urban", "Water quality", "Wetness"), vpu = "01")}}

\item{vpu}{NHDPlusV2 Vector Processing Unit (i.e. Region)}

\item{files}{NHDPlusV2 data file names. Default are used for SNT functions.
See \code{details}}
}
\value{
StreamCat data files are downloaded to "StreamCat" directory
}
\description{
Downloads zip files for VPU from StreamCat webpage
\url{https://www.epa.gov/national-aquatic-resource-surveys/streamcat}
}
\examples{
strmcat_download(strmcat_path = getwd(), Class = c("Disturbance"), Scale =
c("Watershed", "Riparian", "Stream segment"), Metric.Type =
c("Infrastructure"), vpu = "01")

}
