% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_calc.r
\name{net_calc}
\alias{net_calc}
\title{Network Topology Metrics}
\usage{
net_calc(netdelin, vpu, nhdplus_path)
}
\arguments{
\item{netdelin}{output from \code{net_delin}}

\item{vpu}{NHDPlusV2 Vector Processing Unit}

\item{nhdplus_path}{Directory for NHDPlusV2 files (\code{\link{net_nhdplus}})}
}
\value{
\code{data.frame}: \code{$group.comid} stream network root COMID;
  \code{$vpu} NHDPlusV2 vector processing unit;\code{M} Position of sampling
  point on COMID, as proportion of COMID from upstream end; \code{WS.ord}
  strahler order for root node;\code{$head.h2o} number of headwater reaches;
  \code{$trib.jun} number of tributary junctions; \code{reach.cnt} number of
  reaches in network; \code{diver.cnt} count of divergent flow paths;
  \code{$AREASQKM} drainage area (km^2); \code{$LENGTHKM} total lenght of
  network flowlines (km); \code{drain.den} drainage density (\code{LENGTHKM}
  / \code{AREASQKM})
}
\description{
Calculates stream network topology metrics
}
\details{
Requires /NHDPlusAttributes directory (see \code{\link{net_nhdplus}})

Length and area measures are scaled by M values
}
\examples{
# identify NHDPlusV2 COMID
a <- net_sample(nhdplus_path = getwd(), vpu = "01", ws_order = 6, n = 5)
# delineate stream network
b <- net_delin(group_comid = as.character(a[,"COMID"]), nhdplus_path = getwd(), vpu = "01")
calculate topology summary
c <- net_calc(netdelin = b, vpu = "01", nhdplus_path = getwd())
}
