% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_cat.r
\name{net_cat}
\alias{net_cat}
\title{Identify NHDPlus Catchment}
\usage{
net_cat(netdelin, vpu, nhdplus_path, dissolve)
}
\arguments{
\item{netdelin}{output form \code{link{net_delin}}}

\item{vpu}{the vector processing unit}

\item{nhdplus_path}{the directory containing NHDPlus data see(\code{net_nhdplus})}

\item{dissolve}{collapse subcatchments to create a single network catchmet \code{(Y/N)}}
}
\value{
named list: \code{$sf_obj} is simple features object and \code{$na_cat} records
  comid's without catchment
}
\description{
Extracts NHDPlus flowline catchments with optional dissolve
}
\details{
Ploygons can be used to associate landscape attributes not included with
NHDPlus

Requires NHDPlusCatchment directory

Some comid's do not have matching catchment (featureID) and some catchments
do not have matching comid's (see
\url{http://www.horizon-systems.com/NHDPlus/NHDPlusV2_documentation.php})
}
\examples{
d <- net_cat(netdelin = c, vpu = "01", nhdplus_path = getwd(), dissolve = "Y")
plot(st_geometry(d$sf_obj))
}
