% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_comid.r
\name{net_comid}
\alias{net_comid}
\title{Identify Nearest NHDPlus COMID}
\usage{
net_comid(sample_points, CRS, nhdplus_path, vpu, maxdist)
}
\arguments{
\item{sample_points}{\code{data.frame} with field names "SITE_ID", "X" and "Y" to
identify location}

\item{CRS}{coordinate reference system of pts (see examples)}

\item{nhdplus_path}{directory for downloaded NHDPlus files}

\item{vpu}{NHDPlus vector processing unit}

\item{maxdist}{search radius around points (m)}
}
\value{
\code{data.frame} with site identifier (i.e \code{SITE_ID}, \code{X},
  and \code{Y}), x, y coordinates (\code{snap_x,snap_y}) for position on
  NHDPlus flowline (COMID), distance (\code{snap_dist}) in meters to
  \code{pts} argument. \code{GNIS_NAME}, \code{TOTDASQKM} and
  \code{STREAMORDE} are NHDPlus Value Added Attributes and should be used as an aide
  to confirm COMID is correct.
}
\description{
Determines the closest NHDPlus COMID within a radius of a sampling point
}
\details{
NHDSnapshot and NHDPlusAttributes is required NHDlus download (see \code{\link{net_nhdplus}})
}
\examples{
net_comid(sample_points = z, CRS = 4269, nhdplus_path = getwd(), vpu="01", maxdist = 100)

CRS examples: \\code{CRS = 4269} = NAD83 (see
\\url{https://epsg.io/4269});\\code{CRS = 5070} = NAD83/CONUS Albers (see
\\url{https://epsg.io/5070})
}
