% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_comid.r
\name{net_comid}
\alias{net_comid}
\title{Identify NHDPlusV2 COMID}
\usage{
net_comid(sample_points, CRS, nhdplus_path, vpu, maxdist)
}
\arguments{
\item{sample_points}{location \code{data.frame}. fields must be ordered and
named as "SITE_ID", "X" and "Y"}

\item{CRS}{coordinate reference system of locations (see Details)}

\item{nhdplus_path}{directory for NHDPlusV2 files}

\item{vpu}{NHDPlusV2 vector Processing Unit}

\item{maxdist}{search radius around points (m)}
}
\value{
\code{data.frame} with site information (i.e \code{$SITE_ID},
  \code{$X}, and \code{$Y}) and \code{$snap_dist} distance (m) to nearest
  COMID; \code{$snap_x, $snap_y}, coordinates of nearest COMID vertex;
  \code{$M} Position of sampling point on COMID, as proportion of COMID from
  upstream end; \code{COMID} common identifer of NHDPlusV2 flowline;
  \code{GNIS_NAME} Geographic Names Information System name of COMID;
  \code{ApproxTOTDASQKM} drainage area at COMID outlet, may overestimate
  drainage area if \code{$M} < 1; and \code{STREAMORDE} Stream order form
  NHDPlusV2 Value Added Attributes.
}
\description{
Identifies NHDPlusV2 COMID closest to sampling point
}
\details{
NHDSnapshot and NHDPlusAttributes are required NHDlusV2 files (see
\code{\link{net_nhdplus}})

CRS examples: \code{CRS = 4269} = NAD83 (see \url{https://epsg.io/4269});
\code{CRS = 5070} = NAD83/CONUS Albers (see \url{https://epsg.io/5070})
}
\examples{
#read example locations from VPU 11
ExLoc <- read.csv("Sample_Locations.csv")
# reorder and rename location data.frame
ExLoc <- ExLoc[,c("SiteName","W","N")]
names(ExLoc) <- c("SITE_ID","X","Y")
#find nearest NHDPlusV2 COMID
sam_pts <- net_comid(sample_points = ExLoc, CRS = 4269,
                     nhdplus_path = getwd(), vpu = 11, maxdist = 200)
}
