% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_conflu.R
\name{net_conflu}
\alias{net_conflu}
\title{Network Confluence Attributes}
\usage{
net_conflu(netdelin, nhdplus_path, vpu)
}
\arguments{
\item{netdelin}{output from \code{net_delin}}

\item{nhdplus_path}{directory containing NHDPlusV2 \code{\link{net_nhdplus}}}

\item{vpu}{vector processing unit}
}
\value{
\code{data.frame} \code{group.comid} as network identifier;
  \code{net.comid} as the reach identifier; \code{vpu} as vector processing
  unit; \code{CRS} is coordinate reference system for \code{X,Y};
  \code{trib_order} order of COMID downstream of confluence;
  \code{area_ratio} ratio of the darinage areas of each reach in a
  confluence (i.e. Triburaty Drainage Area / Mainstem Drainage Area);
  \code{trib_area} the drainage area downstream of confluence;
  \code{junction_num} is a concatenation of of strahler stream orders of
  confluence reaches (e.g 11 is concluence of 2 frist orders, and 23 is a
  second order joining a thrid order); alpha is the avarage tributary
  junction; \code{complex} indicates complex triburaty junction
}
\description{
Calculates metrics related to network confluences see \code{Values}
}
\details{
requires "/NHDPlusAttributes" directory \code{\link{net_nhdplus}}

netdelin projection is transformed to Lambert Conformal Cone (crs = 102004)
to preserve angles

See Benda et al. (2004) for Triburaty : Mainstem Drainage Ratio and Seybold et
al. (2017) for confluence angle calculation
}
\examples{
h <- net_conflu(netdelin = c, vpu = "01", nhdplus_path = getwd())
}
