% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_delin.r
\name{net_delin}
\alias{net_delin}
\title{Network Delineation}
\usage{
net_delin(group_comid, nhdplus_path, vpu)
}
\arguments{
\item{group_comid}{a vector of NHDPlus COMIDs}

\item{nhdplus_path}{nhdplus_path directory for downloaded NHDPlus files
(\code{\link{net_nhdplus}})}

\item{vpu}{vector processing unit}
}
\value{
netdelin, a named list (\code{$Network, $Nested_COMIDs, $SF_Obj})
  which serves as an argeument for subsequent functions in StreamNetworkTools.
  \code{$Network} stores network comids, \code{Nested_COMIDs} stores comids
  which are in the same network, and SF_Obj is a simple feature object of the
  networks.
}
\description{
Delineates stream network upstream of root NHDPlus COMID (see
\code{\link{net_sample}} or \code{\link{net_comid}} or
\url{http://www.horizon-systems.com/NHDPlus/NHDPlusV2_data.php}) via upstream
navigation
}
\details{
Requires /NHDPlusAttributes and NHDSnapshot directories (see
\code{\link{net_nhdplus}})

All input group comids are delineated. If nested comids are unexpected or
unwanted, remove from \code{$SF_Obj} and \code{$Network} before proceeding
}
\examples{
a <- net_comid(sample_points = z, CRS = 4269, nhdplus_path = getwd(), vpu="01", maxdist = 100)
b <- net_delin(group_comid = as.character(a[,"COMID"]), nhdplus_path = getwd(), vpu = "01")
write_sf(b$SF_Obj...)
}
