% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_hor.r
\name{net_hort}
\alias{net_hort}
\title{Network Area, Length and Bifurcation Ratios}
\usage{
net_hort(netdelin, vpu, nhdplus_path)
}
\arguments{
\item{netdelin}{output from \code{net_delin}}

\item{vpu}{vector processing unit}

\item{nhdplus_path}{directory containing NHDPlusV2 \code{\link{net_nhdplus}}}
}
\value{
named \code{list}: \code{$topology} is \code{data.frame} with count
  of stream numbers (\code{str_num}), sum of stream lengths (\code{str_len})
  and sum of drainage areas (\code{str_area}); and \code{$Horton_Est} are
  horotn estimates \code{Rb,Rl,Ra} (bifurcation, lenght and area ratio,
  respectively). \code{Rb.rsqr, Rl.rsqr, Ra.rsqr} are R^2 values for the
  \code{lm(...)} model. \code{ohm} is the strahler order of the (network -
  1) (see \code{Details}.
}
\description{
Uses \code{lm(...)} to estimates number, length and area ratios for
a stream network (i.e. Horton Laws).
}
\details{
requires /NHDPlusAttributes directory (see \code{net_nhdplus})

\code{ohm} is the strahler order of the (network - 1) because sampling point
(or COMID) my not necessarily encorpporate the entire edge of the largest
order
}
\examples{
net_hort(netdelin = c, vpu = "01", nhdplus_path = getwd())
}
