% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_hor.r
\name{net_hort}
\alias{net_hort}
\title{Horton Laws}
\usage{
net_hort(netdelin, vpu, nhdplus_path)
}
\arguments{
\item{netdelin}{output from \code{net_delin}}

\item{vpu}{vector processing unit}

\item{nhdplus_path}{Directory for NHDPlusV2 files (\code{\link{net_nhdplus}}}
}
\value{
named \code{list}: \code{$topology} is data used to estimate horton
  laws: \code{$group.comid} root COMID for steam network; \code{$str_ord}
  stream order; \code{$str_num} count of stream reaches,\code{str_len} Mean
  reach length; \code{str_area} mean drainage area. \code{$horton_est} are
  estimated horotn ratios: \code{ohm} is the strahler order of the network -
  1 (see \code{Details}. \code{$Rb,$Rl,$Ra} are bifurcation, length and area
  ratios, respectively. \code{Rb.r2, Rl.r2, Ra.r2} are R^2 values from
  \code{lm(...)}.
}
\description{
calculates network area, length and bifurcation ratios, using (\code{lm(...)})
}
\details{
requires /NHDPlusAttributes directory (see \code{net_nhdplus})

Horton estimates (\code{$Horton_est}) include network - 1 (i.e \code{$ohm})
because a given group.comid, may not include the whole stream order
}
\examples{
# identify NHDPlusV2 COMID
a <- net_sample(nhdplus_path = getwd(), vpu = "01", ws_order = 6, n = 5)
# delineate stream network
b <- net_delin(group_comid = as.character(a[,"COMID"]), nhdplus_path = getwd(), vpu = "01")
#estimate horton ratios
c <- net_hort(netdelin = b, vpu = "01", nhdplus_path = getwd())
}
