% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_lc.r
\name{net_lc}
\alias{net_lc}
\title{NLCD2011 Landcover Percentages}
\usage{
net_lc(netdelin, vpu, nhdplus_path)
}
\arguments{
\item{netdelin}{output from \code{\link{net_delin}}}

\item{vpu}{vector processing unit}

\item{nhdplus_path}{directory containing NHDPlus
see(\code{\link{net_nhdplus}})}
}
\value{
\code{data.frame} with NLCD2011 classification codes. Dee
  \url{https://www.mrlc.gov/nlcd11_leg.php} for definitions
}
\description{
Network scale landcover percentages
}
\details{
Requires VPUAttributeExtension directory see(\code{\link{net_nhdplus}})
}
\examples{
net_lc(netdelin = c, vpu = "01", nhdplus_path = getwd())
}
