% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_nhdplus.r
\name{net_nhdplus}
\alias{net_nhdplus}
\title{NHDPlus VPU Download}
\usage{
net_nhdplus(nhdplus_path = getwd(), download = "http", vpu = "01",
  files = "ALL", zip_7 = "C:/Program Files/7-Zip")
}
\arguments{
\item{nhdplus_path}{character specfiying the directory for downloaded files}

\item{download}{character specifying "http" or "ftp"}

\item{vpu}{character vector processing unit for data}

\item{files}{character specifying which files to download. "ALL" will
download full dataset. See Details}

\item{zip_7}{character specifying the location of the 7-zip download}
}
\value{
/NHDPlus and subdirectories for NHDPlus V2 data
}
\description{
\code{net_nhdplus} downloads and extracts the .7z file from the national
hydrography dataset plus by the vector processing unit specified in its
arguments
}
\details{
7-zip \url{https://www.7-zip.org/download.html} must be downloaded and
installed for tool to work

\code{dir} creates a new sub-directory "/NHDPlus" to extract data.

acceptable options for \code{file}: c("FdrFac", "FdrNull", "FilledAreas",
"Hydrodem", "NEDSnapshot", "EROMExtension", "NHDPlusAttributes",
"NHDPlusBurnComponents", "NHDPlusCatchment", "NHDSnapshotFGDB",
"NHDSnapshot", "VogelExtension", "VPUAttributeExtension", "WBDSnapshot"). see
\url{http://www.horizon-systems.com/NHDPlus/NHDPlusV2_documentation.php} for
definitions

All raster rocessing units (RPU) within the vpu will be downloaded

Includes \code{\link[base]{system}} commands which may not work outside
windows OS
}
\examples{
files needed for StreamNetworkTools
net_nhdplus(nhdplus_path=getwd(), download = "http", vpu = "01", files =
c("NHDPlusAttributes", "NHDSnapshot", "NHDPlusCatchment", "VPUAttributeExtension",
"VogelExtension", "EROMExtension"), zip_7 = "C:/Program Files/7-Zip")
}
