% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_sample.r
\name{net_sample}
\alias{net_sample}
\title{Random NHDPlusV2 COMID}
\usage{
net_sample(nhdplus_path, vpu, ws_order, n)
}
\arguments{
\item{nhdplus_path}{directory for NHDPlusV2 files \code{\link{net_nhdplus}}}

\item{vpu}{NHDPlus V2 vector processing unit}

\item{ws_order}{Strahler order of desired COMID}

\item{n}{number of COMID's selected}
}
\value{
vector of NHDPlusV2 COMID's.
}
\description{
Generates random COMID vector of a specified Strahler order. Can serve as
root of steam network (see \code{\link{net_delin}}).
}
\details{
\code{"./NHDPlus"} must be a subdirectory of \code{nhdplus_path}.

NHDPlusAttributes and NHDSnapshot are required NHDlus downloads (see \code{\link{net_nhdplus}})
}
\examples{
net_sample(nhdplus_path=getwd(), vpu = "01", ws_order = 6, n = 5)
}
