% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_sinu.r
\name{net_sinu}
\alias{net_sinu}
\title{Reach Slope and Sinuosity}
\usage{
net_sinu(netdelin, nhdplus_path, vpu)
}
\arguments{
\item{netdelin}{output from \code{net_delin}}

\item{nhdplus_path}{directory containing NHDPlusV2 \code{\link{net_nhdplus}}}

\item{vpu}{vector processing unit}
}
\value{
\code{data.frame} with \code{group.comid} as network identifier;
  \code{net.comid} as the reach identifier; \code{seg.id} as the valley
  segment identifier; \code{tot.len} is the total length of reach;
  \code{str.len} is straight line lenght between begining and end of reach;
  \code{sinuosity} \code{tot.len} / \code{str.len}; \code{MaxElevSM} maximum
  elevation of reach; \code{MinElevSM} minimum elevtion of reach;
  \code{SlopeNHDPlus} slope of reach
}
\description{
Calculates sinuosity and slope for each network COMID
}
\details{
network scale values (i.e. mean) can be derived with \code{aggregate}
}
\examples{
j<-net_sinu(netdelin = c, vpu = "01", nhdplus_path = getwd())
}
